Add files via upload
Task Description
You have been assigned a task to analyze the earning call transcripts for Alphabet Inc from Q12022 to Q22023. The transcripts are available in PDF format on the companv's website. Your task is to
answer two questons:
What are the main products/business offerings by Alphabet Inc., based on their earning calls transcripts?
i.e., can you infer what are the primary products offerings of a company based on its
earning call transcripts?
What are the key initiatives and performance for YouTube across the 6 quarters (Q12022 to Q22023)?
o i.e., can you extract all the relevant paragraphs about YouTube and organize them by
some human-readable categories so that we can understand the trend across time by category.
Dataset
You will be provided with a set of PDFs containing the earning call transcripts for Alphabet Inc from
Q12022 to Q22023. The transcripts are also available on the following website: https://abc.xyz/investor/
Evaluation
It is advisable to consider this assignment as an MVP and manage your time accordingly. It is recommended that vou allocate no more than 6-8 hours to complete it. The primarv evaluation criteria will focus on your problem-solving approach and the assumptions/tradeoffs you make.
Submission
A sample python project template will be provided. The project template will be using poetry for dependencies management. You should update the pyproject.toml file with additional dependencies
requirea for your sorution.
The solution can be jupyter notebooks or python scripts. Your code should be well-documented and easy to follow.
You should also include a short presentation with NOT more than 5 slides to explain your approach, challenges/limitations, and possible improvements.
Bonus Points:
The project template includes FastAPI code snippets for the following API endpoints:
" /products/"
This API takes in a set of PDF transcripts and returns minimally a list of products.
"/product / youtube/
This API takes in a set of PDF transcripts and returns a set of texts that are related to YouTube and their corresponding categories.

Convert your codes into one or more python functions and complete the codes to provide a working
FastAP demo
If possible, also complete the Dockerfile and build a working Docker image to demonstrate the API
service.